pyproject.toml
[project]
name = "ssh-demo"
version = "0.1.0"
requires-python = ">= 3.13.7"
dependencies = [
    "ibis-framework[duckdb]>=10.8.0",
    "plotnine>=0.15.0",
]
